+++ src/main/java/de/maxhenkel/plane/ClientConfig.java
@@ -12,12 +12,14 @@ public class ClientConfig extends ConfigBase {
     public final ForgeConfigSpec.BooleanValue showPlaneInfo;
     public final ForgeConfigSpec.DoubleValue planeInfoScale;
     public final ForgeConfigSpec.EnumValue<SpeedType> planeInfoSpeedType;
+    public final ForgeConfigSpec.DoubleValue planeZoom;
 
     public ClientConfig(ForgeConfigSpec.Builder builder) {
         super(builder);
         showPlaneInfo = builder.define("plane_info.enabled", true);
         planeInfoScale = builder.defineInRange("plane_info.scale", 0.75D, 0.1D, 2D);
         planeInfoSpeedType = builder.defineEnum("plane_info.speed_type", SpeedType.KILOMETERS_PER_HOUR);
+        planeZoom = builder.defineInRange("plane.third_person_zoom", 6D, 1D, 20D);
     }
 
     public static enum SpeedType {